Feat | Add a new option for kind-options (#130)
* Add a new option for kind-options - some of the options that can be passed include --image / --config etc.

One of the benefits is that we can use --config to pass our own cluster config for kind and override any settings as needed including image ver / location for kindest node. Tested by running this locally in my env and seems to work but would require somebody more versed with go to double check.

* Make kindOptions message optional

Co-authored-by: Dag Andersen <dagbjerreandersen@gmail.com>

---------

Co-authored-by: Dag Andersen <dagbjerreandersen@gmail.com>